build(package): downgrade to firebase 8.6.5
chore(itemsFilter): to its own file on util folder
