+++ package.json
@@ -17,7 +17,7 @@
 		"@react-native-community/masked-view": "^0.1.11",
 		"@react-navigation/native": "^5.9.4",
 		"@react-navigation/stack": "^5.14.5",
-		"firebase": "^8.6.8",
+		"firebase": "^8.6.5",
 		"react": "17.0.1",
 		"react-native": "0.64.1",
 		"react-native-dotenv": "^2.5.5",
